% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbBuildNetblockTable}
\alias{nbBuildNetblockTable}
\title{Load netblock data into a table. Returns a pointer to the table.}
\usage{
nbBuildNetblockTable(BaseAndMask, IPaddrStrings, Mask, Description)
}
\description{
Load netblock data into a table. Returns a pointer to the table.
}
