% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netblockr.R
\name{nbLoadNetwork}
\alias{nbLoadNetwork}
\title{Build a network description table from a data frame.}
\usage{
nbLoadNetwork(nets, quiet = FALSE)
}
\arguments{
\item{nets_df}{A data frame with columns named 'NetBlock' & 'Description'}
}
\value{
An external pointer to the table data structure in C++ space
The input can contain empty rows. A '#' in the NetBlock field will cause that row to be ignored.
}
\description{
Build a network description table from a data frame.
}
