% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netblockr.R
\name{nbReadAndLoadNetwork}
\alias{nbReadAndLoadNetwork}
\title{Read a file with netblocks in CIDR format and a free form dscription
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs}
\usage{
nbReadAndLoadNetwork(file, skip_lines = 0)
}
\arguments{
\item{file}{Path to the input file}

\item{skip_lines}{Optional number of lines to skip, use if there is an uncommented header}
}
\value{
An external pointer to the table data structure in C++ space
}
\description{
Read a file with netblocks in CIDR format and a free form dscription
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs
}
\examples{

\dontrun{
organization_net_file <- 'data/org-net.txt'
nbPtrOrg <- nbReadAndLoadNetwork(organization_net_file)

nb <- nbGetNetblockTable(nbPtrOrg)
nb

testAddrs <- c('10.10.10.1', '10.20.10.18', '10.16.3.28', '10.16.8.50', '10.16.9.50',
               '10.16.18.18', '10.16.18.35', '10.48.17.32', '10.50.17.32')


nbLookupIPaddrs(nbPtrOrg, testAddrs)
}

}
