% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netblockr.R
\name{nbReadAndValidate}
\alias{nbReadAndValidate}
\title{Validate a network description file}
\usage{
nbReadAndValidate(file, skip_lines = 0, quiet = FALSE)
}
\arguments{
\item{file}{Path to the input file}

\item{skip_lines}{Optional number of lines to skip, use if there is an uncommented header}
}
\value{
A list of data frames, validatedNets and bad_data
The input can contain empty rows. A '#' in the NetBlock field will cause that row to be ignored.
}
\description{
Note that there is currently no check for duplicate NetBlocks, that can be done externally
Avoid a spectacular crash due to missing netmask
}
