% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nbSetMaskOrder}
\alias{nbSetMaskOrder}
\title{Add the unique netmasks to the table in the desired search order, usually largest netmask first to find shortest match}
\usage{
nbSetMaskOrder(nbt, Masks)
}
\description{
Add the unique netmasks to the table in the desired search order, usually largest netmask first to find shortest match
}
