% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netblockr.R
\name{nbReadAndLoadNetwork}
\alias{nbReadAndLoadNetwork}
\title{Read a file with netblocks in CIDR format and a free form description
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs}
\usage{
nbReadAndLoadNetwork(file, skip_lines = 0, quiet = FALSE)
}
\arguments{
\item{file}{Path to the input file}

\item{skip_lines}{Optional number of lines to skip, use if there is an uncommented header}
}
\value{
An external pointer to the table data structure in C++ space
}
\description{
Read a file with netblocks in CIDR format and a free form description
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs
}
