% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netblockr.R
\name{nbReadAndLoadNetwork}
\alias{nbReadAndLoadNetwork}
\title{Read a file with netblocks in CIDR format and a free form dscription
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs}
\usage{
nbReadAndLoadNetwork(file, skip_lines = 0)
}
\arguments{
\item{file}{Path to the input file}

\item{skip_lines}{Optional number of lines to skip, use if there is an uncommented header}
}
\value{
An external pointer to the table data structure in C++ space
}
\description{
Read a file with netblocks in CIDR format and a free form dscription
of the netblock, then build a table of the network with an index
for fast lookup of IPv4 addresses with nbLookupIPaddrs
}
\examples{
Input file sample (CIDR format netblock and free form description):

# Organization network example  (a comment line)
#  empty lines are ignored and can be used for human readability
   Netblock  Region Code More info (indented lines are comments)

10.16.0.0/12    NOAM xxx North America Supernet
10.16.3.0/24    NOAM PTN Princeton NJ Data Center Servers
10.16.4.0/24    NOAM PTN Princeton NJ Data Center Network
10.16.5.0/24    NOAM PTN Princeton NJ Data Center Management
10.16.6.0/23    NOAM PTN Princeton NJ West Wing Floor #1 '#' in data
10.16.8.0/23    NOAM PTN Princeton NJ West Wing Floor #2
10.16.10.0/23   NOAM PTN Princeton NJ Administration Building
10.16.15.0/24   NOAM PTN Princeton NJ Environmental Controls
10.16.18.0/28   NOAM PTN Princeton NJ VPN Routers
10.16.18.16/28  NOAM PTN Princeton NJ DMZ
10.18.10.0/23   NOAM TOL Tolchester Beach MD
10.18.12.0/23   NOAM SCV Sarah Creek VA

10.32.0.0/12    SOAM xxx South America Supernet
10.33.1.0/20    SOAM RIO Brazil

10.48.0.0/12    EMEA xxx EMEA Supernet
10.48.10.0/23   EMEA LBS London Berkeley Square
10.48.12.0/23   EMEA PSS Portsmouth Southsea
10.48.14.0/23   EMEA IOW Cowes Isle of Wight
10.48.16.0/23   EMEA ZUR Zürich Wasserschöpfi

10.64.0.0/12    APAC xxx APAC Supernet
10.64.10.0/23   APAC SNG Singapore
10.64.12.0/23   APAC TOK Tokyo Heiwajima

Load network data:

 organization_net_file <- 'data/org-net.txt'
 nbPtrOrg <- nbReadAndLoadNetwork(organization_net_file)

}
